requirements.txt
python-telegram-bot[webhooks] @ git+https://github.com/python-telegram-bot/python-telegram-bot.git
together
google-auth
google-auth-oauthlib
google-api-python-client
python-dotenv
asyncpg
starlette
uvicorn